GenAI/pyproject.toml
[project]
name = "GenAI"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "docopt",
    "pypng",
    "seaborn",
    "numpy",
    "pandas",
    "matplotlib",
    "protobuf",
    "h5py",
    "tqdm",
    "joblib",
    "jupyterlab",
    "scipy>=1.16.1",
    "tensorflow-metal>=1.2.0",
    "tensorflow>=2.18.1",
    "tensorflow-datasets>=4.9.9",
    "scikit-learn>=1.7.2",
    "torch>=2.7.0",
    "torchvision>=0.22.1",
    "pytorch-lightning>=2.5.2",
    "click>=8.3.0",
    "huggingface-hub>=1.0.1",
    "music21>=9.9.1",
    "gradio>=6.0.2",
    "pillow>=11.3.0",
]

[tool.uv.sources]
torch = [
  { index = "pytorch-cu128", marker = "sys_platform == 'linux'" },
]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

